[OGLEngine] Fixed issue where big meshes wouldn't render properly. The problem was that I was using unsigned shorts, which are limited to 0-65535. I switched to unsigned int in Mesh, ObjReader and vboindexer classes, and it seems to work when doing a new serialization of .obj files. So keep that in mind, you need to do new serializations of .obj files now.

+++ OGLEngine/OGLEngineGL/Mesh.h
@@ -14,7 +14,7 @@ public:
 	void Draw(Shader* shader);
 	void ApplyTexture(Texture* texture);
 
-	std::vector<unsigned short> indices;
+	std::vector<unsigned int> indices;
 	std::vector<glm::vec3> indexed_vertices;
 	std::vector<glm::vec2> indexed_uvs;
 	std::vector<glm::vec3> indexed_normals;
